ADC readings are now averaged to get more stable results but display is slower. Also removed decimal place on display

+++ main/ADS1230/ADS1230.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 
 void ADS1230_init();
-float ADS1230_read();
+int32_t ADS1230_read();
 bool hasData();
 void calibrateADC();
 void calibrateWeight();
